+++ engines/got/views/game_content.cpp
@@ -207,6 +207,14 @@ bool GameContent::tick() {
 		break;
 	}
 
+	if (_G(eyeballs) == 1) { // eyeballs movement animation
+		if (!_G(setup).f25) {
+			_G(thor)->_dir = 0;
+		} else {
+			_G(thor)->_dir = 1;
+		}
+	}
+
 	checkForAreaChange();
 
 	// Check for end of game area
@@ -255,7 +263,7 @@ void GameContent::drawActors(GfxSurface &s) {
 
 	for (int actor_num = 0; actor_num <= MAX_ACTORS;) {
 		// Check for blinking flag
-		if (actor_ptr->_active && !(actor_ptr->_show & 2)) {
+		if (actor_ptr && actor_ptr->_active && !(actor_ptr->_show & 2)) {
 			actor_ptr->_lastX[_G(pge)] = actor_ptr->_x;
 			actor_ptr->_lastY[_G(pge)] = actor_ptr->_y;
 
@@ -499,8 +507,10 @@ void GameContent::thorDies() {
 void GameContent::spinThor() {
 	static const byte DIRS[] = {0, 2, 1, 3};
 
-	_G(thor)->_dir = DIRS[(_deathCtr / SPIN_INTERVAL) % 4];
-	_G(thor)->_lastDir = DIRS[(_deathCtr / SPIN_INTERVAL) % 4];
+	if (!_G(eyeballs)) {
+		_G(thor)->_dir = DIRS[(_deathCtr / SPIN_INTERVAL) % 4];
+		_G(thor)->_lastDir = DIRS[(_deathCtr / SPIN_INTERVAL) % 4];
+	}
 
 	++_deathCtr;
 }