+++ engines/scumm/scumm.cpp
@@ -201,6 +201,13 @@ ScummEngine::ScummEngine(OSystem *syst, const DetectorResult &dr)
 		}
 	}
 
+	if (_game.platform == Common::kPlatformSegaCD) {
+		ConfMan.registerDefault("enable_sega_shadow_mode", false);
+		if (ConfMan.hasKey("enable_sega_shadow_mode", _targetName)) {
+			_enableSegaShadowMode = ConfMan.getBool("enable_sega_shadow_mode");
+		}
+	}
+
 	if (ConfMan.hasKey("gamma_correction", _targetName)) {
 		_useGammaCorrection = ConfMan.getBool("gamma_correction");
 	}